requirements.txt
streamlit==0.62.1
pytesseract==0.3.6
numpy==1.18.5
pandas==1.0.3
opencv_python==4.2.0.32
Pillow==8.0.1
